Rename components/ScrollableChart.vue to components/index/_shared/ScrollableChart.vue
