+++ components/index/CardsMonitoring/HospitalizedNumber/Chart.vue
@@ -69,7 +69,7 @@ import DataViewTable, {
   TableItem,
 } from '@/components/DataViewTable.vue'
 import OpenDataLink from '@/components/index/_shared/OpenDataLink.vue'
-import ScrollableChart from '@/components/ScrollableChart.vue'
+import ScrollableChart from '@/components/index/_shared/ScrollableChart.vue'
 import {
   DataSets,
   DataSetsPoint,
