+++ components/index/CardsMonitoring/PositiveRate/Chart.vue
@@ -113,7 +113,7 @@ import DataViewTable, {
   TableItem,
 } from '@/components/DataViewTable.vue'
 import OpenDataLink from '@/components/index/_shared/OpenDataLink.vue'
-import ScrollableChart from '@/components/ScrollableChart.vue'
+import ScrollableChart from '@/components/index/_shared/ScrollableChart.vue'
 import {
   DisplayData,
   yAxesBgPlugin,
