+++ components/index/_shared/ScrollableChart.vue
